requirements.txt

greenlet
gevent
eventlet
aiohttp
aiohttp_session
aiohttp-devtools
pyyaml
cryptography
aiohttp_cors
graphql_ws
graphene
lxml
cryptography
anytree
asyncio
aiofiles
git+https://github.com/FreeOpcUa/opcua-asyncio.git
